LPS-114698 Rename to match interface package
